# Backend/internal/user/dto.go
package user

type RegisterReq struct {
    Email    string `json:"email"`
    Password string `json:"password"`
    Name     string `json:"name"`
}

type LoginReq struct {
    Email    string `json:"email"`
    Password string `json:"password"`
}

type UserResponse struct {
    ID    string `json:"id"`
    Name  string `json:"name"`
    Email string `json:"email"`
    Role  string `json:"role"`
}

type RefreshReq struct {
	RefreshToken string `json:"refreshtoken"`
}

type LogoutReq struct {
	RefreshID string `json:"refreshId"`
}